refactor(workspace-store): migrate AsyncAPI schemas to v3.0 structure
- Updated import paths for AsyncAPI document types to reflect the new v3.0 schema organization.
- Removed deprecated AsyncAPI schema files and replaced them with new v3.0 schema definitions, enhancing clarity and maintainability.
- Ensured all references to AsyncAPI documents across the codebase are consistent with the new structure, improving type safety and organization.
