+++ packages/workspace-store/src/client.ts
@@ -20,11 +20,11 @@ import { isAsyncApiDocument, isOpenApiDocument } from '@/helpers/type-guards'
 import { createAsyncApiNavigation, createNavigation } from '@/navigation'
 import { externalValueResolver, loadingStatus, refsEverywhere, restoreOriginalRefs } from '@/plugins'
 import { getServersFromDocument } from '@/preprocessing/server'
-import { type AsyncApiDocument, AsyncApiDocumentSchemaStrict } from '@/schemas/asyncapi'
 import { extensions } from '@/schemas/extensions'
 import { type InMemoryWorkspace, InMemoryWorkspaceSchema } from '@/schemas/inmemory-workspace'
 import { defaultReferenceConfig } from '@/schemas/reference-config'
 import { coerceValue } from '@/schemas/typebox-coerce'
+import { type AsyncApiDocument, AsyncApiDocumentSchemaStrict } from '@/schemas/v3.0'
 import {
   OpenAPIDocumentSchema as OpenAPIDocumentSchemaStrict,
   type OpenApiDocument,
