+++ packages/workspace-store/src/helpers/type-guards.ts
@@ -1,6 +1,6 @@
 import type { UnknownObject } from '@/helpers/general'
 import type { ApiDefinition } from '@/schemas'
-import type { AsyncApiDocument } from '@/schemas/asyncapi'
+import type { AsyncApiDocument } from '@/schemas/v3.0'
 import type { OpenApiDocument } from '@/schemas/v3.1/strict/openapi-document'
 
 export function isOpenApiDocument(doc: ApiDefinition | UnknownObject): doc is OpenApiDocument {
