+++ packages/workspace-store/src/mutators/helpers.ts
@@ -1,5 +1,5 @@
 import type { WorkspaceStore } from '@/client'
-import type { AsyncApiDocument } from '@/schemas/asyncapi/asyncapi-document'
+import type { AsyncApiDocument } from '@/schemas/v3.0/asyncapi-document'
 import type { OpenApiDocument } from '@/schemas/v3.1/strict/openapi-document'
 
 export const getDocument = (
