+++ packages/workspace-store/src/schemas/asyncapi/v3.0/operation-trait.ts
