+++ packages/workspace-store/src/schemas/asyncapi/v3.0/server.ts
